--- src/store/loaded/index.ts
@@ -1,1 +1,0 @@
-export * from './loaded';
