--- src/store/todos/index.ts
@@ -1,1 +1,0 @@
-export * from './todos';
